+++ CHANGELOG.md
@@ -5,6 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.92.0] - 2024-03-01 02:01:53
+
+### Added
+
+- Nova Scotia disability amount.
+
+## [0.91.0] - 2024-02-29 14:47:21
+
+### Added
+
+- British Columbia pension income amount.
+
+## [0.90.0] - 2024-02-29 02:49:05
+
+### Added
+
+- British Columbia basic personal amount.
+
 ## [0.89.0] - 2024-02-24 05:08:23
 
 ### Added
@@ -700,6 +718,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
+[0.92.0]: https://github.com/PolicyEngine/policyengine-canada/compare/0.91.0...0.92.0
+[0.91.0]: https://github.com/PolicyEngine/policyengine-canada/compare/0.90.0...0.91.0
+[0.90.0]: https://github.com/PolicyEngine/policyengine-canada/compare/0.89.0...0.90.0
 [0.89.0]: https://github.com/PolicyEngine/policyengine-canada/compare/0.88.1...0.89.0
 [0.88.1]: https://github.com/PolicyEngine/policyengine-canada/compare/0.88.0...0.88.1
 [0.88.0]: https://github.com/PolicyEngine/policyengine-canada/compare/0.87.1...0.88.0